Split submitters out into their own modules.
