+++ python/ray/data/_internal/execution/operators/map_operator_state.py
@@ -1,8 +1,9 @@
-from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import Callable, Optional, List, Dict, Any, Union, Tuple, Iterator
 
 import ray
+from ray.data.block import Block, BlockMetadata
+from ray.data.context import DatasetContext
 from ray.data._internal.compute import (
     ComputeStrategy,
     TaskPoolStrategy,
@@ -12,11 +13,12 @@ from ray.data._internal.execution.util import merge_ref_bundles
 from ray.data._internal.execution.interfaces import (
     RefBundle,
 )
-from ray.data._internal.delegating_block_builder import DelegatingBlockBuilder
+from ray.data._internal.execution.operators.map_task_submitter import MapTaskSubmitter
+from ray.data._internal.execution.operators.actor_pool_submitter import (
+    ActorPoolSubmitter,
+)
+from ray.data._internal.execution.operators.task_pool_submitter import TaskPoolSubmitter
 from ray.data._internal.memory_tracing import trace_allocation
-from ray.data._internal.remote_fn import cached_remote_fn
-from ray.data.block import Block, BlockAccessor, BlockMetadata, BlockExecStats
-from ray.data.context import DEFAULT_SCHEDULING_STRATEGY, DatasetContext
 from ray.types import ObjectRef
 from ray._raylet import ObjectRefGenerator
 
@@ -224,283 +226,3 @@ class _TaskState:
     output: Optional[RefBundle] = None
     #  TODO(Clark): Remove this once dynamic block splitting is supported for actors.
     block_metadata_ref: Optional[ObjectRef[BlockMetadata]] = None
-
-
-class MapTaskSubmitter(ABC):
-    """A task submitter for MapOperator.
-
-    This abstraction is in charge of submitting tasks, reserving resources for their
-    execution, and cleaning up said resources when a task completes or when task
-    submission is done.
-    """
-
-    def start(self):
-        """Start the task submitter so it's ready to submit tasks.
-
-        This is called when execution of the map operator actually starts, and is where
-        the submitter can initialize expensive state, reserve resources, start workers,
-        etc.
-        """
-        pass
-
-    @abstractmethod
-    def submit(
-        self,
-        transform_fn: ObjectRef[Callable[[Iterator[Block]], Iterator[Block]]],
-        input_blocks: List[ObjectRef[Block]],
-    ) -> Union[
-        ObjectRef[ObjectRefGenerator], Tuple[ObjectRef[Block], ObjectRef[BlockMetadata]]
-    ]:
-        """Submit a map task.
-
-        Args:
-            transform_fn: The function to apply to a block bundle in the submitted
-                map task.
-            input_blocks: The block bundle on which to apply transform_fn.
-
-        Returns:
-            An object ref representing the output of the map task.
-        """
-        raise NotImplementedError
-
-    def task_done(self, task_ref: ray.ObjectRef):
-        """Indicates that the task that output the provided ref is done.
-
-        Args:
-            task_ref: The output ref for the task that's done.
-        """
-        pass
-
-    def task_submission_done(self):
-        """Indicates that no more tasks will be submitter."""
-        pass
-
-    @abstractmethod
-    def cancellable(self):
-        """Whether the submitted tasks are cancellable."""
-        raise NotImplementedError
-
-
-class TaskPoolSubmitter(MapTaskSubmitter):
-    """A task submitter for MapOperator that uses normal Ray tasks."""
-
-    def __init__(self, ray_remote_args: Dict[str, Any]):
-        """Create a TaskPoolSubmitter instance.
-
-        Args:
-            ray_remote_args: Remote arguments for the Ray tasks to be launched.
-        """
-        self._ray_remote_args = ray_remote_args
-
-    def submit(
-        self,
-        transform_fn: ObjectRef[Callable[[Iterator[Block]], Iterator[Block]]],
-        input_blocks: List[ObjectRef[Block]],
-    ) -> ObjectRef[ObjectRefGenerator]:
-        # Submit the task as a normal Ray task.
-        map_task = cached_remote_fn(_map_task, num_returns="dynamic")
-        return map_task.options(**self._ray_remote_args).remote(
-            transform_fn, *input_blocks
-        )
-
-    def cancellable(self):
-        # Normal Ray tasks are cancellable.
-        return True
-
-
-def _map_task(
-    fn: Callable[[Iterator[Block]], Iterator[Block]],
-    *blocks: Block,
-) -> Iterator[Union[Block, List[BlockMetadata]]]:
-    """Remote function for a single operator task.
-
-    Args:
-        fn: The callable that takes Iterator[Block] as input and returns
-            Iterator[Block] as output.
-        blocks: The concrete block values from the task ref bundle.
-
-    Returns:
-        A generator of blocks, followed by the list of BlockMetadata for the blocks
-        as the last generator return.
-    """
-    output_metadata = []
-    stats = BlockExecStats.builder()
-    for b_out in fn(iter(blocks)):
-        # TODO(Clark): Add input file propagation from input blocks.
-        m_out = BlockAccessor.for_block(b_out).get_metadata([], None)
-        m_out.exec_stats = stats.build()
-        output_metadata.append(m_out)
-        yield b_out
-        stats = BlockExecStats.builder()
-    yield output_metadata
-
-
-class ActorPoolSubmitter(MapTaskSubmitter):
-    """A task submitter for MapOperator that uses a Ray actor pool."""
-
-    def __init__(
-        self, compute_strategy: ActorPoolStrategy, ray_remote_args: Dict[str, Any]
-    ):
-        """Create an ActorPoolSubmitter instance.
-
-        Args:
-            compute_strategy: The configured ActorPoolStrategy.
-            ray_remote_args: Remote arguments for the Ray actors to be created.
-        """
-        # TODO(Clark): Better mapping from configured min/max pool size to static pool
-        # size?
-        pool_size = compute_strategy.max_size
-        if pool_size == float("inf"):
-            pool_size = compute_strategy.min_size
-        self._pool_size = pool_size
-        self._ray_remote_args = ray_remote_args
-        # A map from task output futures to the actors on which they are running.
-        self._active_actors: Dict[ObjectRef[Block], ray.actor.ActorHandle] = {}
-        # The actor pool on which we are running map tasks.
-        self._actor_pool = ActorPool()
-
-    def start(self):
-        # Create the actor workers and add them to the pool.
-        ray_remote_args = self._apply_default_remote_args(self._ray_remote_args)
-        cls_ = ray.remote(**ray_remote_args)(MapWorker)
-        for _ in range(self._pool_size):
-            self._actor_pool.add_actor(cls_.remote())
-
-    def submit(
-        self,
-        transform_fn: ObjectRef[Callable[[Iterator[Block]], Iterator[Block]]],
-        input_blocks: List[ObjectRef[Block]],
-    ) -> Tuple[ObjectRef[Block], ObjectRef[BlockMetadata]]:
-        # Pick an actor from the pool.
-        actor = self._actor_pool.pick_actor()
-        # Submit the map task.
-        block, block_metadata = actor.submit.options(num_returns=2).remote(
-            transform_fn, *input_blocks
-        )
-        self._active_actors[block] = actor
-        return block, block_metadata
-
-    def task_done(self, ref: ObjectRef[Block]):
-        # Return the actor that was running the task to the pool.
-        actor = self._active_actors.pop(ref)
-        self._actor_pool.return_actor(actor)
-
-    def task_submission_done(self):
-        # Kill all idle actors in the pool, and ensure that all remaining actors in the
-        # pool will be killed as they become idle.
-        self._actor_pool.kill_idle_actors()
-        self._actor_pool.kill_future_idle_actors()
-
-    def cancellable(self):
-        # Actor tasks are not cancellable.
-        return False
-
-    @staticmethod
-    def _apply_default_remote_args(ray_remote_args: Dict[str, Any]) -> Dict[str, Any]:
-        """Apply defaults to the actor creation remote args."""
-        ray_remote_args = ray_remote_args.copy()
-        if "num_cpus" not in ray_remote_args:
-            ray_remote_args["num_cpus"] = 1
-        if "scheduling_strategy" not in ray_remote_args:
-            ctx = DatasetContext.get_current()
-            if ctx.scheduling_strategy == DEFAULT_SCHEDULING_STRATEGY:
-                ray_remote_args["scheduling_strategy"] = "SPREAD"
-            else:
-                ray_remote_args["scheduling_strategy"] = ctx.scheduling_strategy
-        return ray_remote_args
-
-
-class MapWorker:
-    """An actor worker for MapOperator."""
-
-    def ready(self):
-        return "ok"
-
-    def submit(
-        self, fn: Callable[[Iterator[Block]], Iterator[Block]], *blocks: Block
-    ) -> Tuple[Block, BlockMetadata]:
-        # Coalesce all fn output blocks.
-        # TODO(Clark): Remove this coalescing once dynamic block splitting is supported
-        # for actors.
-        # yield from _map_task(fn, *blocks)
-        stats = BlockExecStats.builder()
-        builder = DelegatingBlockBuilder()
-        for block in fn(iter(blocks)):
-            builder.add_block(block)
-        block = builder.build()
-        block_metadata = BlockAccessor.for_block(block).get_metadata([], None)
-        block_metadata.exec_stats = stats.build()
-        return block, block_metadata
-
-
-class ActorPool:
-    """A pool of actors for map task execution.
-
-    This class is in charge of tracking the number of in-flight tasks per actor,
-    providing the least heavily loaded actor to the task submitter, and killing idle
-    actors when the task submitter is done submitting work to the pool.
-    """
-
-    def __init__(self):
-        # Number of tasks in flight per actor.
-        self._tasks_in_flight: Dict[ray.actor.ActorHandle, int] = {}
-        # Whether actors that become idle should be eagerly killed. This is False until
-        # the first call to kill_idle_actors().
-        self._should_kill_idle_actors = False
-
-    def add_actor(self, actor: ray.actor.ActorHandle):
-        """Adds an actor to the pool."""
-        self._tasks_in_flight[actor] = 0
-
-    def pick_actor(self) -> ray.actor.ActorHandle:
-        """Provides the least heavily loaded actor in the pool for task submission."""
-        actor = min(
-            self._tasks_in_flight.keys(), key=lambda actor: self._tasks_in_flight[actor]
-        )
-        self._tasks_in_flight[actor] += 1
-        return actor
-
-    def return_actor(self, actor: ray.actor.ActorHandle):
-        """Returns the provided actor to the pool."""
-        self._tasks_in_flight[actor] -= 1
-        if self._should_kill_idle_actors and self._tasks_in_flight[actor] == 0:
-            self._kill_actor(actor)
-
-    def has_actor(self, actor: ray.actor.ActorHandle) -> bool:
-        """Whether the provided actor is in this pool."""
-        return actor in self._tasks_in_flight
-
-    def get_tasks_in_flight(self, actor: ray.actor.ActorHandle) -> int:
-        """The number of tasks in flight for the provided actor."""
-        return self._tasks_in_flight[actor]
-
-    @property
-    def num_actors(self) -> int:
-        """The number of actors in the pool."""
-        return len(self._tasks_in_flight)
-
-    def kill_idle_actors(self):
-        """Kills all currently idle actors, and ensures that all actors that become idle
-        in the future will be eagerly killed.
-
-        This is called once the task submitter is done submitting work to the pool.
-        """
-        idle_actors = [
-            actor
-            for actor, tasks_in_flight in self._tasks_in_flight.items()
-            if tasks_in_flight == 0
-        ]
-        for actor in idle_actors:
-            self._kill_actor(actor)
-
-    def kill_future_idle_actors(self):
-        """Ensure that all actors that become idle in the future will be eagerly killed.
-
-        This is called once the task submitter is done submitting work to the pool.
-        """
-        self._should_kill_idle_actors = True
-
-    def _kill_actor(self, actor: ray.actor.ActorHandle):
-        """Kill the provided actor and remove it from the pool."""
-        ray.kill(actor)
-        del self._tasks_in_flight[actor]
